Middleware/mcp-supabase/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
pydantic
python-dotenv
httpx
